@@ -6,25 +6,21 @@
 
 #include "camodocal/camera_models/Camera.h"
 
-namespace camodocal
-{
+namespace camodocal {
 
-class CameraFactory
-{
-public:
-    EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-    CameraFactory();
+class CameraFactory {
+ public:
+   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
+   CameraFactory();
 
-    static boost::shared_ptr<CameraFactory> instance(void);
+   static boost::shared_ptr<CameraFactory> instance();
 
-    CameraPtr generateCamera(Camera::ModelType modelType,
-                             const std::string& cameraName,
-                             cv::Size imageSize) const;
+   static CameraPtr generateCamera(Camera::ModelType modelType, const std::string& cameraName, const cv::Size& imageSize) ;
 
-    CameraPtr generateCameraFromYamlFile(const std::string& filename);
+   static CameraPtr generateCameraFromYamlFile(const std::string& filename);
 
-private:
-    static boost::shared_ptr<CameraFactory> m_instance;
+ private:
+   static boost::shared_ptr<CameraFactory> m_instance;
 };
 
 }
