@@ -15,70 +15,56 @@ typedef boost::shared_ptr<ChessboardQuad> ChessboardQuadPtr;
 
 class Chessboard
 {
-public:
-    Chessboard(cv::Size boardSize, cv::Mat& image);
+ public:
+   Chessboard(cv::Size boardSize, cv::Mat& image);
 
-    void findCorners(bool useOpenCV = false);
-    const std::vector<cv::Point2f>& getCorners(void) const;
-    bool cornersFound(void) const;
+   void findCorners(bool useOpenCV = false);
+   const std::vector<cv::Point2f>& getCorners() const;
+   bool cornersFound() const;
 
-    const cv::Mat& getImage(void) const;
-    const cv::Mat& getSketch(void) const;
+   const cv::Mat& getImage() const;
+   const cv::Mat& getSketch() const;
 
-private:
-    bool findChessboardCorners(const cv::Mat& image,
-                               const cv::Size& patternSize,
-                               std::vector<cv::Point2f>& corners,
-                               int flags, bool useOpenCV);
+ private:
+   bool findChessboardCorners(const cv::Mat& image,const cv::Size& patternSize,
+                              std::vector<cv::Point2f>& corners, int flags, bool useOpenCV);
 
-    bool findChessboardCornersImproved(const cv::Mat& image,
-                                       const cv::Size& patternSize,
-                                       std::vector<cv::Point2f>& corners,
-                                       int flags);
+   bool findChessboardCornersImproved(const cv::Mat& image, const cv::Size& patternSize,
+                                      std::vector<cv::Point2f>& corners, int flags);
 
-    void cleanFoundConnectedQuads(std::vector<ChessboardQuadPtr>& quadGroup, cv::Size patternSize);
+   void cleanFoundConnectedQuads(std::vector<ChessboardQuadPtr>& quadGroup, cv::Size patternSize);
 
-    void findConnectedQuads(std::vector<ChessboardQuadPtr>& quads,
-                            std::vector<ChessboardQuadPtr>& group,
-                            int group_idx, int dilation);
+   void findConnectedQuads(std::vector<ChessboardQuadPtr>& quads, std::vector<ChessboardQuadPtr>& group, int group_idx, int dilation);
 
 //    int checkQuadGroup(std::vector<ChessboardQuadPtr>& quadGroup,
 //                       std::vector<ChessboardCornerPtr>& outCorners,
 //                       cv::Size patternSize);
 
-    void labelQuadGroup(std::vector<ChessboardQuadPtr>& quad_group,
-                        cv::Size patternSize, bool firstRun);
+   void labelQuadGroup(std::vector<ChessboardQuadPtr>& quad_group, cv::Size patternSize, bool firstRun);
 
-    void findQuadNeighbors(std::vector<ChessboardQuadPtr>& quads, int dilation);
+   void findQuadNeighbors(std::vector<ChessboardQuadPtr>& quads, int dilation);
 
-    int augmentBestRun(std::vector<ChessboardQuadPtr>& candidateQuads, int candidateDilation,
-                       std::vector<ChessboardQuadPtr>& existingQuads, int existingDilation);
+   int augmentBestRun(std::vector<ChessboardQuadPtr>& candidateQuads, int candidateDilation,
+                      std::vector<ChessboardQuadPtr>& existingQuads, int existingDilation);
 
-    void generateQuads(std::vector<ChessboardQuadPtr>& quads,
-                       cv::Mat& image, int flags,
-                       int dilation, bool firstRun);
+   void generateQuads(std::vector<ChessboardQuadPtr>& quads, cv::Mat& image, int flags, int dilation, bool firstRun);
 
-    bool checkQuadGroup(std::vector<ChessboardQuadPtr>& quads,
-                        std::vector<ChessboardCornerPtr>& corners,
-                        cv::Size patternSize);
+   bool checkQuadGroup(std::vector<ChessboardQuadPtr>& quads, std::vector<ChessboardCornerPtr>& corners, cv::Size patternSize);
 
-    void getQuadrangleHypotheses(const std::vector< std::vector<cv::Point> >& contours,
-                                 std::vector< std::pair<float, int> >& quads,
-                                 int classId) const;
+   void getQuadrangleHypotheses(const std::vector< std::vector<cv::Point> >& contours,
+                                std::vector< std::pair<float, int> >& quads, int classId) const;
 
-    bool checkChessboard(const cv::Mat& image, cv::Size patternSize) const;
+   bool checkChessboard(const cv::Mat& image, const cv::Size& patternSize) const;
 
-    bool checkBoardMonotony(std::vector<ChessboardCornerPtr>& corners,
-                            cv::Size patternSize);
+   bool checkBoardMonotony(std::vector<ChessboardCornerPtr>& corners, cv::Size patternSize);
 
-    bool matchCorners(ChessboardQuadPtr& quad1, int corner1,
-                      ChessboardQuadPtr& quad2, int corner2) const;
+   bool matchCorners(ChessboardQuadPtr& quad1, int corner1, ChessboardQuadPtr& quad2, int corner2) const;
 
-    cv::Mat mImage;
-    cv::Mat mSketch;
-    std::vector<cv::Point2f> mCorners;
-    cv::Size mBoardSize;
-    bool mCornersFound;
+   cv::Mat mImage; // 待检测棋盘格角点的灰度图像
+   cv::Mat mSketch; // 棋盘格角点检测结果的彩色图像
+   std::vector<cv::Point2f> mCorners; // 检测到的棋盘格角点
+   cv::Size mBoardSize; // 棋盘格的尺寸（Size包含两个整型的成员变量width和height）
+   bool mCornersFound; // 是否检测到棋盘格角点
 };
 
 }
