@@ -3,34 +3,34 @@
 
 #include <boost/shared_ptr.hpp>
 #include <eigen3/Eigen/Dense>
-#include <stdint.h>
+#include <cstdint>
 
 namespace camodocal
 {
 
 class Transform
 {
-public:
-    EIGEN_MAKE_ALIGNED_OPERATOR_NEW
+ public:
+   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 
-    Transform();
-    Transform(const Eigen::Matrix4d& H);
+   Transform();
+   explicit Transform(const Eigen::Matrix4d& H);
 
-    Eigen::Quaterniond& rotation(void);
-    const Eigen::Quaterniond& rotation(void) const;
-    double* rotationData(void);
-    const double* const rotationData(void) const;
+   Eigen::Quaterniond& rotation();
+   const Eigen::Quaterniond& rotation() const;
+   double* rotationData();
+   const double* rotationData() const;
 
-    Eigen::Vector3d& translation(void);
-    const Eigen::Vector3d& translation(void) const;
-    double* translationData(void);
-    const double* const translationData(void) const;
+   Eigen::Vector3d& translation();
+   const Eigen::Vector3d& translation() const;
+   double* translationData();
+   const double* translationData() const;
 
-    Eigen::Matrix4d toMatrix(void) const;
+   Eigen::Matrix4d toMatrix() const;
 
-private:
-    Eigen::Quaterniond m_q;
-    Eigen::Vector3d m_t;
+ private:
+   Eigen::Quaterniond m_q;
+   Eigen::Vector3d m_t;
 };
 
 }
