@@ -20,46 +20,46 @@ using namespace std;
 using namespace camodocal;
 using namespace Eigen;
 
-bool inBorder(const cv::Point2f &pt);
+bool inBorder(const cv::Point2f& pt);
 
-void reduceVector(vector<cv::Point2f> &v, vector<uchar> status);
-void reduceVector(vector<int> &v, vector<uchar> status);
+void reduceVector(vector<cv::Point2f>& v, vector<uchar> status);
+void reduceVector(vector<int>& v, vector<uchar> status);
 
 class FeatureTracker
 {
-  public:
-    FeatureTracker();
+ public:
+   FeatureTracker();
 
-    void readImage(const cv::Mat &_img,double _cur_time);
+   void readImage(const cv::Mat& _img, double _cur_time);
 
-    void setMask();
+   void setMask();
 
-    void addPoints();
+   void addPoints();
 
-    bool updateID(unsigned int i);
+   bool updateID(unsigned int i);
 
-    void readIntrinsicParameter(const string &calib_file);
+   void readIntrinsicParameter(const string& calib_file);
 
-    void showUndistortion(const string &name);
+   void showUndistortion(const string& name);
 
-    void rejectWithF();
+   void rejectWithF();
 
-    void undistortedPoints();
+   void undistortedPoints();
 
-    cv::Mat mask;
-    cv::Mat fisheye_mask;
-    cv::Mat prev_img, cur_img, forw_img;
-    vector<cv::Point2f> n_pts;
-    vector<cv::Point2f> prev_pts, cur_pts, forw_pts;
-    vector<cv::Point2f> prev_un_pts, cur_un_pts;
-    vector<cv::Point2f> pts_velocity;
-    vector<int> ids;
-    vector<int> track_cnt;
-    map<int, cv::Point2f> cur_un_pts_map;
-    map<int, cv::Point2f> prev_un_pts_map;
-    camodocal::CameraPtr m_camera;
-    double cur_time;
-    double prev_time;
+   cv::Mat mask;
+   cv::Mat fisheye_mask;
+   cv::Mat prev_img, cur_img, forw_img;
+   vector<cv::Point2f> n_pts; // 新特征点
+   vector<cv::Point2f> prev_pts, cur_pts, forw_pts;
+   vector<cv::Point2f> prev_un_pts, cur_un_pts;
+   vector<cv::Point2f> pts_velocity;
+   vector<int> ids;
+   vector<int> track_cnt;
+   map<int, cv::Point2f> cur_un_pts_map;
+   map<int, cv::Point2f> prev_un_pts_map;
+   camodocal::CameraPtr m_camera; // 各种相机类型的基类
+   double cur_time{}; // 当前帧图像的时间戳
+   double prev_time{}; // 上一图像帧的时间戳
 
-    static int n_id;
+   static int n_id;
 };
