@@ -1,71 +1,77 @@
 #include "marginalization_factor.h"
 
+/**
+ * @brief 待边缘化的各个参数块计算残差和雅可比，同时处理核函数
+ */
 void ResidualBlockInfo::Evaluate()
 {
-    residuals.resize(cost_function->num_residuals());
+  residuals.resize(cost_function->num_residuals()); // 残差的维度
 
-    std::vector<int> block_sizes = cost_function->parameter_block_sizes();
-    raw_jacobians = new double *[block_sizes.size()];
-    jacobians.resize(block_sizes.size());
+  std::vector<int> block_sizes = cost_function->parameter_block_sizes(); // 确定相关参数块的数目
+  raw_jacobians = new double* [block_sizes.size()]; // ceres接口都是double数组，因此这里给雅可比准备数组
+  jacobians.resize(block_sizes.size());
 
-    for (int i = 0; i < static_cast<int>(block_sizes.size()); i++)
+  // 这里把jacobians的每个矩阵地址赋给raw_jacobians，然后把raw_jacobians传递给ceres的接口，这样计算结果就会直接写入jacobians中
+  for (int i = 0; i < static_cast<int>(block_sizes.size()); i++)
+  {
+    jacobians[i].resize(cost_function->num_residuals(), block_sizes[i]);
+    raw_jacobians[i] = jacobians[i].data(); // 取出指针
+    //dim += block_sizes[i] == 7 ? 6 : block_sizes[i];
+  }
+  // 调用各自重载的Evaluate函数，计算残差和雅可比
+  cost_function->Evaluate(parameter_blocks.data(), residuals.data(), raw_jacobians);
+
+  //std::vector<int> tmp_idx(block_sizes.size());
+  //Eigen::MatrixXd tmp(dim, dim);
+  //for (int i = 0; i < static_cast<int>(parameter_blocks.size()); i++)
+  //{
+  //    int size_i = localSize(block_sizes[i]);
+  //    Eigen::MatrixXd jacobian_i = jacobians[i].leftCols(size_i);
+  //    for (int j = 0, sub_idx = 0; j < static_cast<int>(parameter_blocks.size()); sub_idx += block_sizes[j] == 7 ? 6 : block_sizes[j], j++)
+  //    {
+  //        int size_j = localSize(block_sizes[j]);
+  //        Eigen::MatrixXd jacobian_j = jacobians[j].leftCols(size_j);
+  //        tmp_idx[j] = sub_idx;
+  //        tmp.block(tmp_idx[i], tmp_idx[j], size_i, size_j) = jacobian_i.transpose() * jacobian_j;
+  //    }
+  //}
+  //Eigen::SelfAdjointEigenSolver<Eigen::MatrixXd> saes(tmp);
+  //std::cout << saes.eigenvalues() << std::endl;
+  //ROS_ASSERT(saes.eigenvalues().minCoeff() >= -1e-6);
+
+  if (loss_function)
+  {
+    double residual_scaling_, alpha_sq_norm_;
+
+    double sq_norm, rho[3];
+
+    sq_norm = residuals.squaredNorm(); // 残差的模
+    loss_function->Evaluate(sq_norm, rho);  // pho[0]是核函数值，pho[1]是核函数的一阶导数，pho[2]是核函数的二阶导数
+    //printf("sq_norm: %f, rho[0]: %f, rho[1]: %f, rho[2]: %f\n", sq_norm, rho[0], rho[1], rho[2]);
+
+    double sqrt_rho1_ = sqrt(rho[1]);
+
+    if ((sq_norm == 0.0) || (rho[2] <= 0.0)) //
     {
-        jacobians[i].resize(cost_function->num_residuals(), block_sizes[i]);
-        raw_jacobians[i] = jacobians[i].data();
-        //dim += block_sizes[i] == 7 ? 6 : block_sizes[i];
+      residual_scaling_ = sqrt_rho1_;
+      alpha_sq_norm_ = 0.0;
     }
-    cost_function->Evaluate(parameter_blocks.data(), residuals.data(), raw_jacobians);
-
-    //std::vector<int> tmp_idx(block_sizes.size());
-    //Eigen::MatrixXd tmp(dim, dim);
-    //for (int i = 0; i < static_cast<int>(parameter_blocks.size()); i++)
-    //{
-    //    int size_i = localSize(block_sizes[i]);
-    //    Eigen::MatrixXd jacobian_i = jacobians[i].leftCols(size_i);
-    //    for (int j = 0, sub_idx = 0; j < static_cast<int>(parameter_blocks.size()); sub_idx += block_sizes[j] == 7 ? 6 : block_sizes[j], j++)
-    //    {
-    //        int size_j = localSize(block_sizes[j]);
-    //        Eigen::MatrixXd jacobian_j = jacobians[j].leftCols(size_j);
-    //        tmp_idx[j] = sub_idx;
-    //        tmp.block(tmp_idx[i], tmp_idx[j], size_i, size_j) = jacobian_i.transpose() * jacobian_j;
-    //    }
-    //}
-    //Eigen::SelfAdjointEigenSolver<Eigen::MatrixXd> saes(tmp);
-    //std::cout << saes.eigenvalues() << std::endl;
-    //ROS_ASSERT(saes.eigenvalues().minCoeff() >= -1e-6);
-
-    if (loss_function)
+    else
     {
-        double residual_scaling_, alpha_sq_norm_;
-
-        double sq_norm, rho[3];
-
-        sq_norm = residuals.squaredNorm();
-        loss_function->Evaluate(sq_norm, rho);
-        //printf("sq_norm: %f, rho[0]: %f, rho[1]: %f, rho[2]: %f\n", sq_norm, rho[0], rho[1], rho[2]);
-
-        double sqrt_rho1_ = sqrt(rho[1]);
-
-        if ((sq_norm == 0.0) || (rho[2] <= 0.0))
-        {
-            residual_scaling_ = sqrt_rho1_;
-            alpha_sq_norm_ = 0.0;
-        }
-        else
-        {
-            const double D = 1.0 + 2.0 * sq_norm * rho[2] / rho[1];
-            const double alpha = 1.0 - sqrt(D);
-            residual_scaling_ = sqrt_rho1_ / (1 - alpha);
-            alpha_sq_norm_ = alpha / sq_norm;
-        }
-
-        for (int i = 0; i < static_cast<int>(parameter_blocks.size()); i++)
-        {
-            jacobians[i] = sqrt_rho1_ * (jacobians[i] - alpha_sq_norm_ * residuals * (residuals.transpose() * jacobians[i]));
-        }
-
-        residuals *= residual_scaling_;
+      const double D = 1.0 + 2.0 * sq_norm * rho[2] / rho[1];
+      const double alpha = 1.0 - sqrt(D);
+      residual_scaling_ = sqrt_rho1_ / (1 - alpha);
+      alpha_sq_norm_ = alpha / sq_norm;
     }
+
+    // 这里相当于残差和雅可比都乘以sqrt_rho1_，即核函数所在点的一阶导数
+    for (int i = 0; i < static_cast<int>(parameter_blocks.size()); i++)
+    {
+      jacobians[i] = sqrt_rho1_ * (jacobians[i] - alpha_sq_norm_ * residuals * (residuals.transpose() * jacobians[i]));
+    }
+
+    residuals *= residual_scaling_;
+  }
 }
 
 MarginalizationInfo::~MarginalizationInfo()
@@ -86,51 +92,62 @@ MarginalizationInfo::~MarginalizationInfo()
     }
 }
 
-void MarginalizationInfo::addResidualBlockInfo(ResidualBlockInfo *residual_block_info)
+/**
+ * @brief 收集各个残差
+ * @param[in] residual_block_info
+ */
+void MarginalizationInfo::addResidualBlockInfo(ResidualBlockInfo* residual_block_info)
 {
-    factors.emplace_back(residual_block_info);
+  factors.emplace_back(residual_block_info); // 将ResidualBlockInfo指针放入factors中
 
-    std::vector<double *> &parameter_blocks = residual_block_info->parameter_blocks;
-    std::vector<int> parameter_block_sizes = residual_block_info->cost_function->parameter_block_sizes();
+  std::vector<double*>& parameter_blocks = residual_block_info->parameter_blocks; // 这个是和该约束相关的参数块
+  std::vector<int> parameter_block_sizes = residual_block_info->cost_function->parameter_block_sizes(); // 各个参数块的大小
 
-    for (int i = 0; i < static_cast<int>(residual_block_info->parameter_blocks.size()); i++)
-    {
-        double *addr = parameter_blocks[i];
-        int size = parameter_block_sizes[i];
-        parameter_block_size[reinterpret_cast<long>(addr)] = size;
-    }
+  for (int i = 0; i < static_cast<int>(residual_block_info->parameter_blocks.size()); i++)
+  {
+    double* addr = parameter_blocks[i];
+    int size = parameter_block_sizes[i];
+    // 这里是个unordered_map，避免重复添加
+    parameter_block_size[reinterpret_cast<long>(addr)] = size;
+  }
 
-    for (int i = 0; i < static_cast<int>(residual_block_info->drop_set.size()); i++)
-    {
-        double *addr = parameter_blocks[residual_block_info->drop_set[i]];
-        parameter_block_idx[reinterpret_cast<long>(addr)] = 0;
-    }
+  // 待边缘化的参数块
+  for (int i: residual_block_info->drop_set)
+  {
+    double* addr = parameter_blocks[i];
+    // 先准备好待边缘化的参数块的unordered_map
+    parameter_block_idx[reinterpret_cast<long>(addr)] = 0;
+  }
 }
 
+/**
+ * @brief 将各个残差块计算残差和雅克比，同时备份所有相关的参数块内容
+ */
 void MarginalizationInfo::preMarginalize()
 {
-    for (auto it : factors)
-    {
-        it->Evaluate();
+  for (auto it: factors) // 遍历所有的残差块
+  {
+    it->Evaluate(); // 调用这个接口计算各个残差块的残差和雅克比
 
-        std::vector<int> block_sizes = it->cost_function->parameter_block_sizes();
-        for (int i = 0; i < static_cast<int>(block_sizes.size()); i++)
-        {
-            long addr = reinterpret_cast<long>(it->parameter_blocks[i]);
-            int size = block_sizes[i];
-            if (parameter_block_data.find(addr) == parameter_block_data.end())
-            {
-                double *data = new double[size];
-                memcpy(data, it->parameter_blocks[i], sizeof(double) * size);
-                parameter_block_data[addr] = data;
-            }
-        }
+    std::vector<int> block_sizes = it->cost_function->parameter_block_sizes(); // 得到每个残差块的参数块大小
+    for (int i = 0; i < static_cast<int>(block_sizes.size()); i++)
+    {
+      long addr = reinterpret_cast<long>(it->parameter_blocks[i]); // 每一个参数块的地址
+      int size = block_sizes[i];
+      // 把每一个参数块都备份起来，使用unordered_map避免参数块重复，之所以备份，是为了后面的状态保留
+      if (parameter_block_data.find(addr) == parameter_block_data.end())
+      {
+        auto* data = new double[size];
+        memcpy(data, it->parameter_blocks[i], sizeof(double) * size);
+        parameter_block_data[addr] = data; // 地址->参数块实际内容的地址
+      }
     }
+  }
 }
 
-int MarginalizationInfo::localSize(int size) const
+int MarginalizationInfo::localSize(int size)
 {
-    return size == 7 ? 6 : size;
+  return size == 7 ? 6 : size;
 }
 
 int MarginalizationInfo::globalSize(int size) const
@@ -140,242 +157,269 @@ int MarginalizationInfo::globalSize(int size) const
 
 void* ThreadsConstructA(void* threadsstruct)
 {
-    ThreadsStruct* p = ((ThreadsStruct*)threadsstruct);
-    for (auto it : p->sub_factors)
+  ThreadsStruct* p = ((ThreadsStruct*)threadsstruct);
+  // 遍历分配过来的任务
+  for (auto it: p->sub_factors)
+  {
+    for (int i = 0; i < static_cast<int>(it->parameter_blocks.size()); i++)
     {
-        for (int i = 0; i < static_cast<int>(it->parameter_blocks.size()); i++)
-        {
-            int idx_i = p->parameter_block_idx[reinterpret_cast<long>(it->parameter_blocks[i])];
-            int size_i = p->parameter_block_size[reinterpret_cast<long>(it->parameter_blocks[i])];
-            if (size_i == 7)
-                size_i = 6;
-            Eigen::MatrixXd jacobian_i = it->jacobians[i].leftCols(size_i);
-            for (int j = i; j < static_cast<int>(it->parameter_blocks.size()); j++)
-            {
-                int idx_j = p->parameter_block_idx[reinterpret_cast<long>(it->parameter_blocks[j])];
-                int size_j = p->parameter_block_size[reinterpret_cast<long>(it->parameter_blocks[j])];
-                if (size_j == 7)
-                    size_j = 6;
-                Eigen::MatrixXd jacobian_j = it->jacobians[j].leftCols(size_j);
-                if (i == j)
-                    p->A.block(idx_i, idx_j, size_i, size_j) += jacobian_i.transpose() * jacobian_j;
-                else
-                {
-                    p->A.block(idx_i, idx_j, size_i, size_j) += jacobian_i.transpose() * jacobian_j;
-                    p->A.block(idx_j, idx_i, size_j, size_i) = p->A.block(idx_i, idx_j, size_i, size_j).transpose();
-                }
-            }
-            p->b.segment(idx_i, size_i) += jacobian_i.transpose() * it->residuals;
-        }
-    }
-    return threadsstruct;
-}
-
-void MarginalizationInfo::marginalize()
-{
-    int pos = 0;
-    for (auto &it : parameter_block_idx)
-    {
-        it.second = pos;
-        pos += localSize(parameter_block_size[it.first]);
-    }
-
-    m = pos;
-
-    for (const auto &it : parameter_block_size)
-    {
-        if (parameter_block_idx.find(it.first) == parameter_block_idx.end())
-        {
-            parameter_block_idx[it.first] = pos;
-            pos += localSize(it.second);
-        }
-    }
-
-    n = pos - m;
-
-    //ROS_DEBUG("marginalization, pos: %d, m: %d, n: %d, size: %d", pos, m, n, (int)parameter_block_idx.size());
-
-    TicToc t_summing;
-    Eigen::MatrixXd A(pos, pos);
-    Eigen::VectorXd b(pos);
-    A.setZero();
-    b.setZero();
-    /*
-    for (auto it : factors)
-    {
-        for (int i = 0; i < static_cast<int>(it->parameter_blocks.size()); i++)
-        {
-            int idx_i = parameter_block_idx[reinterpret_cast<long>(it->parameter_blocks[i])];
-            int size_i = localSize(parameter_block_size[reinterpret_cast<long>(it->parameter_blocks[i])]);
-            Eigen::MatrixXd jacobian_i = it->jacobians[i].leftCols(size_i);
-            for (int j = i; j < static_cast<int>(it->parameter_blocks.size()); j++)
-            {
-                int idx_j = parameter_block_idx[reinterpret_cast<long>(it->parameter_blocks[j])];
-                int size_j = localSize(parameter_block_size[reinterpret_cast<long>(it->parameter_blocks[j])]);
-                Eigen::MatrixXd jacobian_j = it->jacobians[j].leftCols(size_j);
-                if (i == j)
-                    A.block(idx_i, idx_j, size_i, size_j) += jacobian_i.transpose() * jacobian_j;
-                else
-                {
-                    A.block(idx_i, idx_j, size_i, size_j) += jacobian_i.transpose() * jacobian_j;
-                    A.block(idx_j, idx_i, size_j, size_i) = A.block(idx_i, idx_j, size_i, size_j).transpose();
-                }
-            }
-            b.segment(idx_i, size_i) += jacobian_i.transpose() * it->residuals;
-        }
-    }
-    ROS_INFO("summing up costs %f ms", t_summing.toc());
-    */
-    //multi thread
-
-
-    TicToc t_thread_summing;
-    pthread_t tids[NUM_THREADS];
-    ThreadsStruct threadsstruct[NUM_THREADS];
-    int i = 0;
-    for (auto it : factors)
-    {
-        threadsstruct[i].sub_factors.push_back(it);
-        i++;
-        i = i % NUM_THREADS;
-    }
-    for (int i = 0; i < NUM_THREADS; i++)
-    {
-        TicToc zero_matrix;
-        threadsstruct[i].A = Eigen::MatrixXd::Zero(pos,pos);
-        threadsstruct[i].b = Eigen::VectorXd::Zero(pos);
-        threadsstruct[i].parameter_block_size = parameter_block_size;
-        threadsstruct[i].parameter_block_idx = parameter_block_idx;
-        int ret = pthread_create( &tids[i], NULL, ThreadsConstructA ,(void*)&(threadsstruct[i]));
-        if (ret != 0)
-        {
-            ROS_WARN("pthread_create error");
-            ROS_BREAK();
-        }
-    }
-    for( int i = NUM_THREADS - 1; i >= 0; i--)  
-    {
-        pthread_join( tids[i], NULL ); 
-        A += threadsstruct[i].A;
-        b += threadsstruct[i].b;
-    }
-    //ROS_DEBUG("thread summing up costs %f ms", t_thread_summing.toc());
-    //ROS_INFO("A diff %f , b diff %f ", (A - tmp_A).sum(), (b - tmp_b).sum());
-
-
-    //TODO
-    Eigen::MatrixXd Amm = 0.5 * (A.block(0, 0, m, m) + A.block(0, 0, m, m).transpose());
-    Eigen::SelfAdjointEigenSolver<Eigen::MatrixXd> saes(Amm);
-
-    //ROS_ASSERT_MSG(saes.eigenvalues().minCoeff() >= -1e-4, "min eigenvalue %f", saes.eigenvalues().minCoeff());
-
-    Eigen::MatrixXd Amm_inv = saes.eigenvectors() * Eigen::VectorXd((saes.eigenvalues().array() > eps).select(saes.eigenvalues().array().inverse(), 0)).asDiagonal() * saes.eigenvectors().transpose();
-    //printf("error1: %f\n", (Amm * Amm_inv - Eigen::MatrixXd::Identity(m, m)).sum());
-
-    Eigen::VectorXd bmm = b.segment(0, m);
-    Eigen::MatrixXd Amr = A.block(0, m, m, n);
-    Eigen::MatrixXd Arm = A.block(m, 0, n, m);
-    Eigen::MatrixXd Arr = A.block(m, m, n, n);
-    Eigen::VectorXd brr = b.segment(m, n);
-    A = Arr - Arm * Amm_inv * Amr;
-    b = brr - Arm * Amm_inv * bmm;
-
-    Eigen::SelfAdjointEigenSolver<Eigen::MatrixXd> saes2(A);
-    Eigen::VectorXd S = Eigen::VectorXd((saes2.eigenvalues().array() > eps).select(saes2.eigenvalues().array(), 0));
-    Eigen::VectorXd S_inv = Eigen::VectorXd((saes2.eigenvalues().array() > eps).select(saes2.eigenvalues().array().inverse(), 0));
-
-    Eigen::VectorXd S_sqrt = S.cwiseSqrt();
-    Eigen::VectorXd S_inv_sqrt = S_inv.cwiseSqrt();
-
-    linearized_jacobians = S_sqrt.asDiagonal() * saes2.eigenvectors().transpose();
-    linearized_residuals = S_inv_sqrt.asDiagonal() * saes2.eigenvectors().transpose() * b;
-    //std::cout << A << std::endl
-    //          << std::endl;
-    //std::cout << linearized_jacobians << std::endl;
-    //printf("error2: %f %f\n", (linearized_jacobians.transpose() * linearized_jacobians - A).sum(),
-    //      (linearized_jacobians.transpose() * linearized_residuals - b).sum());
-}
-
-std::vector<double *> MarginalizationInfo::getParameterBlocks(std::unordered_map<long, double *> &addr_shift)
-{
-    std::vector<double *> keep_block_addr;
-    keep_block_size.clear();
-    keep_block_idx.clear();
-    keep_block_data.clear();
-
-    for (const auto &it : parameter_block_idx)
-    {
-        if (it.second >= m)
-        {
-            keep_block_size.push_back(parameter_block_size[it.first]);
-            keep_block_idx.push_back(parameter_block_idx[it.first]);
-            keep_block_data.push_back(parameter_block_data[it.first]);
-            keep_block_addr.push_back(addr_shift[it.first]);
-        }
-    }
-    sum_block_size = std::accumulate(std::begin(keep_block_size), std::end(keep_block_size), 0);
-
-    return keep_block_addr;
-}
-
-MarginalizationFactor::MarginalizationFactor(MarginalizationInfo* _marginalization_info):marginalization_info(_marginalization_info)
-{
-    int cnt = 0;
-    for (auto it : marginalization_info->keep_block_size)
-    {
-        mutable_parameter_block_sizes()->push_back(it);
-        cnt += it;
-    }
-    //printf("residual size: %d, %d\n", cnt, n);
-    set_num_residuals(marginalization_info->n);
-};
-
-bool MarginalizationFactor::Evaluate(double const *const *parameters, double *residuals, double **jacobians) const
-{
-    //printf("internal addr,%d, %d\n", (int)parameter_block_sizes().size(), num_residuals());
-    //for (int i = 0; i < static_cast<int>(keep_block_size.size()); i++)
-    //{
-    //    //printf("unsigned %x\n", reinterpret_cast<unsigned long>(parameters[i]));
-    //    //printf("signed %x\n", reinterpret_cast<long>(parameters[i]));
-    //printf("jacobian %x\n", reinterpret_cast<long>(jacobians));
-    //printf("residual %x\n", reinterpret_cast<long>(residuals));
-    //}
-    int n = marginalization_info->n;
-    int m = marginalization_info->m;
-    Eigen::VectorXd dx(n);
-    for (int i = 0; i < static_cast<int>(marginalization_info->keep_block_size.size()); i++)
-    {
-        int size = marginalization_info->keep_block_size[i];
-        int idx = marginalization_info->keep_block_idx[i] - m;
-        Eigen::VectorXd x = Eigen::Map<const Eigen::VectorXd>(parameters[i], size);
-        Eigen::VectorXd x0 = Eigen::Map<const Eigen::VectorXd>(marginalization_info->keep_block_data[i], size);
-        if (size != 7)
-            dx.segment(idx, size) = x - x0;
+      int idx_i = p->parameter_block_idx[reinterpret_cast<long>(it->parameter_blocks[i])]; // 在大矩阵A中的索引
+      int size_i = p->parameter_block_size[reinterpret_cast<long>(it->parameter_blocks[i])]; // 参数块的大小
+      if (size_i == 7)
+        size_i = 6;
+      Eigen::MatrixXd jacobian_i = it->jacobians[i].leftCols(size_i); // 四元数是过参数化的，求导是用的是李代数，因此要取左边6列
+      // 和本身以及其他雅可比块构造H矩阵，i是当前参数块，j是其他参数块
+      for (int j = i; j < static_cast<int>(it->parameter_blocks.size()); j++)
+      {
+        int idx_j = p->parameter_block_idx[reinterpret_cast<long>(it->parameter_blocks[j])];
+        int size_j = p->parameter_block_size[reinterpret_cast<long>(it->parameter_blocks[j])];
+        if (size_j == 7)
+          size_j = 6;
+        Eigen::MatrixXd jacobian_j = it->jacobians[j].leftCols(size_j);
+        if (i == j)
+          p->A.block(idx_i, idx_j, size_i, size_j) += jacobian_i.transpose() * jacobian_j;
         else
         {
-            dx.segment<3>(idx + 0) = x.head<3>() - x0.head<3>();
-            dx.segment<3>(idx + 3) = 2.0 * Utility::positify(Eigen::Quaterniond(x0(6), x0(3), x0(4), x0(5)).inverse() * Eigen::Quaterniond(x(6), x(3), x(4), x(5))).vec();
-            if (!((Eigen::Quaterniond(x0(6), x0(3), x0(4), x0(5)).inverse() * Eigen::Quaterniond(x(6), x(3), x(4), x(5))).w() >= 0))
-            {
-                dx.segment<3>(idx + 3) = 2.0 * -Utility::positify(Eigen::Quaterniond(x0(6), x0(3), x0(4), x0(5)).inverse() * Eigen::Quaterniond(x(6), x(3), x(4), x(5))).vec();
-            }
+          p->A.block(idx_i, idx_j, size_i, size_j) += jacobian_i.transpose() * jacobian_j;
+          p->A.block(idx_j, idx_i, size_j, size_i) = p->A.block(idx_i, idx_j, size_i, size_j).transpose();
         }
+      }
+      p->b.segment(idx_i, size_i) += jacobian_i.transpose() * it->residuals; // 正常情况下 g = -JT * e
     }
-    Eigen::Map<Eigen::VectorXd>(residuals, n) = marginalization_info->linearized_residuals + marginalization_info->linearized_jacobians * dx;
-    if (jacobians)
-    {
+  }
+  return threadsstruct;
+}
 
-        for (int i = 0; i < static_cast<int>(marginalization_info->keep_block_size.size()); i++)
-        {
-            if (jacobians[i])
-            {
-                int size = marginalization_info->keep_block_size[i], local_size = marginalization_info->localSize(size);
-                int idx = marginalization_info->keep_block_idx[i] - m;
-                Eigen::Map<Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>> jacobian(jacobians[i], n, size);
-                jacobian.setZero();
-                jacobian.leftCols(local_size) = marginalization_info->linearized_jacobians.middleCols(idx, local_size);
-            }
-        }
+/**
+ * @brief 边缘化处理，并将结果转换成残差和雅可比的形式
+ */
+void MarginalizationInfo::marginalize()
+{
+  int pos = 0;
+  // parameter_block_idx是unordered_map，key是参数块地址，value初始都是0
+  for (auto& it: parameter_block_idx)
+  {
+    it.second = pos; // 这是所有参数块中排序的id，待边缘化的参数块排在前面
+    pos += localSize(parameter_block_size[it.first]); // 因为要求导，所以是localSize，具体一点就是使用李代数
+  }
+
+  m = pos;
+
+  for (const auto& it: parameter_block_size) // 遍历所有的参数块
+  {
+    if (parameter_block_idx.find(it.first) == parameter_block_idx.end())
+    {
+      parameter_block_idx[it.first] = pos;
+      pos += localSize(it.second);
     }
-    return true;
+  }
+
+  n = pos - m;
+
+  //ROS_DEBUG("marginalization, pos: %d, m: %d, n: %d, size: %d", pos, m, n, (int)parameter_block_idx.size());
+
+  TicToc t_summing;
+  Eigen::MatrixXd A(pos, pos); // Ax = b 预设大小
+  Eigen::VectorXd b(pos);
+  A.setZero();
+  b.setZero();
+  /*
+  for (auto it : factors)
+  {
+      for (int i = 0; i < static_cast<int>(it->parameter_blocks.size()); i++)
+      {
+          int idx_i = parameter_block_idx[reinterpret_cast<long>(it->parameter_blocks[i])];
+          int size_i = localSize(parameter_block_size[reinterpret_cast<long>(it->parameter_blocks[i])]);
+          Eigen::MatrixXd jacobian_i = it->jacobians[i].leftCols(size_i);
+          for (int j = i; j < static_cast<int>(it->parameter_blocks.size()); j++)
+          {
+              int idx_j = parameter_block_idx[reinterpret_cast<long>(it->parameter_blocks[j])];
+              int size_j = localSize(parameter_block_size[reinterpret_cast<long>(it->parameter_blocks[j])]);
+              Eigen::MatrixXd jacobian_j = it->jacobians[j].leftCols(size_j);
+              if (i == j)
+                  A.block(idx_i, idx_j, size_i, size_j) += jacobian_i.transpose() * jacobian_j;
+              else
+              {
+                  A.block(idx_i, idx_j, size_i, size_j) += jacobian_i.transpose() * jacobian_j;
+                  A.block(idx_j, idx_i, size_j, size_i) = A.block(idx_i, idx_j, size_i, size_j).transpose();
+              }
+          }
+          b.segment(idx_i, size_i) += jacobian_i.transpose() * it->residuals;
+      }
+  }
+  ROS_INFO("summing up costs %f ms", t_summing.toc());
+  */
+  //multi thread
+
+  // 往矩阵A和向量b中填充数据，利用多线程加速
+  TicToc t_thread_summing;
+  pthread_t tids[NUM_THREADS]; // pthread_t 是 POSIX 线程库中用于标识线程id的数据类型
+  ThreadsStruct threadsstruct[NUM_THREADS]; // 用于传递给线程的结构体数组
+  int i = 0;
+  for (auto it: factors)
+  {
+    threadsstruct[i].sub_factors.push_back(it); // 将残差块分配给不同的线程
+    i++;
+    i = i % NUM_THREADS;
+  }
+  // 每一个线程构造一个A和b，最后再合并
+  for (int i = 0; i < NUM_THREADS; i++)
+  {
+    TicToc zero_matrix;
+    // 所以 矩阵A和向量b的大小都是pos，初始值都是0
+    threadsstruct[i].A = Eigen::MatrixXd::Zero(pos,pos);
+    threadsstruct[i].b = Eigen::VectorXd::Zero(pos);
+    threadsstruct[i].parameter_block_size = parameter_block_size; // 大小
+    threadsstruct[i].parameter_block_idx = parameter_block_idx; // 索引
+    // 当你创建一个新的线程时（例如，使用 pthread_create 函数），一个 pthread_t 类型的变量会被用作线程 ID 的存放位置
+    int ret = pthread_create( &tids[i], nullptr, ThreadsConstructA ,(void*)&(threadsstruct[i]));
+    if (ret != 0)
+    {
+      ROS_WARN("pthread_create error");
+      ROS_BREAK();
+    }
+  }
+  for( int i = NUM_THREADS - 1; i >= 0; i--)
+  {
+    // 等待各个线程结束
+    pthread_join( tids[i], nullptr );
+    A += threadsstruct[i].A;
+    b += threadsstruct[i].b;
+  }
+  //ROS_DEBUG("thread summing up costs %f ms", t_thread_summing.toc());
+  //ROS_INFO("A diff %f , b diff %f ", (A - tmp_A).sum(), (b - tmp_b).sum());
+
+
+  //TODO
+  // Amm矩阵的构建是为了保证正定性
+  Eigen::MatrixXd Amm = 0.5 * (A.block(0, 0, m, m) + A.block(0, 0, m, m).transpose());
+  Eigen::SelfAdjointEigenSolver<Eigen::MatrixXd> saes(Amm); // 实对称矩阵或者复厄米特矩阵的特征值分解
+
+  //ROS_ASSERT_MSG(saes.eigenvalues().minCoeff() >= -1e-4, "min eigenvalue %f", saes.eigenvalues().minCoeff());
+
+  // 一个逆矩阵的特征值是其原矩阵的特征值的倒数，特征向量不变，利用特征值取逆来构造逆矩阵，select类似于C++里的？：
+  Eigen::MatrixXd Amm_inv = saes.eigenvectors() * Eigen::VectorXd((saes.eigenvalues().array() > eps).select(saes.eigenvalues().array().inverse(), 0)).asDiagonal() * saes.eigenvectors().transpose();
+  //printf("error1: %f\n", (Amm * Amm_inv - Eigen::MatrixXd::Identity(m, m)).sum());
+
+  Eigen::VectorXd bmm = b.segment(0, m); // 待边缘化的大小
+  Eigen::MatrixXd Amr = A.block(0, m, m, n); // lambda_b
+  Eigen::MatrixXd Arm = A.block(m, 0, n, m);
+  Eigen::MatrixXd Arr = A.block(m, m, n, n); // lambda_c
+  Eigen::VectorXd brr = b.segment(m, n);
+  A = Arr - Arm * Amm_inv * Amr;
+  b = brr - Arm * Amm_inv * bmm;
+
+  Eigen::SelfAdjointEigenSolver<Eigen::MatrixXd> saes2(A);
+  Eigen::VectorXd S = Eigen::VectorXd((saes2.eigenvalues().array() > eps).select(saes2.eigenvalues().array(), 0));
+  Eigen::VectorXd S_inv = Eigen::VectorXd((saes2.eigenvalues().array() > eps).select(saes2.eigenvalues().array().inverse(), 0));
+
+  Eigen::VectorXd S_sqrt = S.cwiseSqrt(); // 这个求得的是S^0.5，这里是向量
+  Eigen::VectorXd S_inv_sqrt = S_inv.cwiseSqrt();
+
+  linearized_jacobians = S_sqrt.asDiagonal() * saes2.eigenvectors().transpose();
+  linearized_residuals = S_inv_sqrt.asDiagonal() * saes2.eigenvectors().transpose() * b;
+  //std::cout << A << std::endl
+  //          << std::endl;
+  //std::cout << linearized_jacobians << std::endl;
+  //printf("error2: %f %f\n", (linearized_jacobians.transpose() * linearized_jacobians - A).sum(),
+  //      (linearized_jacobians.transpose() * linearized_residuals - b).sum());
+}
+
+std::vector<double*> MarginalizationInfo::getParameterBlocks(std::unordered_map<long, double*>& addr_shift)
+{
+  std::vector<double*> keep_block_addr;
+  keep_block_size.clear();
+  keep_block_idx.clear();
+  keep_block_data.clear();
+
+  for (const auto& it : parameter_block_idx) // 遍历边缘化相关的每个参数块
+  {
+    if (it.second >= m)
+    {
+      keep_block_size.push_back(parameter_block_size[it.first]); // 留下来的参数块的大小 global size
+      keep_block_idx.push_back(parameter_block_idx[it.first]); // 留下来的参数块的索引
+      keep_block_data.push_back(parameter_block_data[it.first]); // 留下来的参数块的内容
+      keep_block_addr.push_back(addr_shift[it.first]); // 对应的新地址
+    }
+  }
+  // 保留的参数块的总大小
+  sum_block_size = std::accumulate(std::begin(keep_block_size), std::end(keep_block_size), 0);
+
+  return keep_block_addr;
+}
+
+/**
+ * @brief 边缘化信息结果的构造函数，根据边缘化信息确定参数块总数和残差维数
+ * @param _marginalization_info
+ */
+MarginalizationFactor::MarginalizationFactor(MarginalizationInfo* _marginalization_info): marginalization_info(_marginalization_info)
+{
+  int cnt = 0;
+  for (auto it: marginalization_info->keep_block_size) // keep_block_size表示上一次边缘化保留的参数块的大小
+  {
+    mutable_parameter_block_sizes()->push_back(it); // 这里是ceres的接口，用于确定参数块的大小
+    cnt += it;
+  }
+  //printf("residual size: %d, %d\n", cnt, n);
+  set_num_residuals(marginalization_info->n); // 这里是ceres的接口，用于确定残差的维度，local size
+};
+
+/**
+ * @brief 边缘化结果的残差和雅可比的计算
+ * @param[in] parameters
+ * @param[out] residuals
+ * @param[out] jacobians
+ * @return bool
+ */
+bool MarginalizationFactor::Evaluate(double const* const* parameters, double* residuals, double** jacobians) const
+{
+  //printf("internal addr,%d, %d\n", (int)parameter_block_sizes().size(), num_residuals());
+  //for (int i = 0; i < static_cast<int>(keep_block_size.size()); i++)
+  //{
+  //    //printf("unsigned %x\n", reinterpret_cast<unsigned long>(parameters[i]));
+  //    //printf("signed %x\n", reinterpret_cast<long>(parameters[i]));
+  //printf("jacobian %x\n", reinterpret_cast<long>(jacobians));
+  //printf("residual %x\n", reinterpret_cast<long>(residuals));
+  //}
+  int n = marginalization_info->n;
+  int m = marginalization_info->m; // 边缘化掉的参数块的大小
+  Eigen::VectorXd dx(n);
+  for (int i = 0; i < static_cast<int>(marginalization_info->keep_block_size.size()); i++)
+  {
+    int size = marginalization_info->keep_block_size[i];
+    int idx = marginalization_info->keep_block_idx[i] - m; // idx起点统一为0
+    Eigen::VectorXd x = Eigen::Map<const Eigen::VectorXd>(parameters[i], size); // 当前参数块的值
+    Eigen::VectorXd x0 = Eigen::Map<const Eigen::VectorXd>(marginalization_info->keep_block_data[i], size); // 上一次边缘化保留的参数块的值
+    if (size != 7)
+      dx.segment(idx, size) = x - x0; // 不需要local parameterization的，直接作差
+    else
+    {
+      dx.segment<3>(idx + 0) = x.head<3>() - x0.head<3>();
+      dx.segment<3>(idx + 3) = 2.0 * Utility::positify(Eigen::Quaterniond(x0(6), x0(3), x0(4), x0(5)).inverse() * Eigen::Quaterniond(x(6), x(3), x(4), x(5))).vec();
+      if ((Eigen::Quaterniond(x0(6), x0(3), x0(4), x0(5)).inverse() *
+           Eigen::Quaterniond(x(6), x(3), x(4), x(5))).w() < 0)
+      {
+        dx.segment<3>(idx + 3) = 2.0 * -Utility::positify(Eigen::Quaterniond(x0(6), x0(3), x0(4), x0(5)).inverse() * Eigen::Quaterniond(x(6), x(3), x(4), x(5))).vec();
+      }
+    }
+  }
+
+  // 更新残差，边缘化后的先验误差 e = e0 + J * dx，根据FEJ，雅可比保持不变，但是残差随着优化会变化，因此下面只更新残差，不更新雅可比
+  // 详情见https://www.zhihu.com/question/52869487、https://blog.csdn.net/weixin_41394379/article/details/89975386
+  Eigen::Map<Eigen::VectorXd>(residuals, n) = marginalization_info->linearized_residuals + marginalization_info->linearized_jacobians * dx;
+  if (jacobians)
+  {
+    for (int i = 0; i < static_cast<int>(marginalization_info->keep_block_size.size()); i++)
+    {
+      if (jacobians[i])
+      {
+        int size = marginalization_info->keep_block_size[i], local_size = marginalization_info->localSize(size);
+        int idx = marginalization_info->keep_block_idx[i] - m;
+        Eigen::Map<Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>> jacobian(jacobians[i], n, size);
+        jacobian.setZero();
+        jacobian.leftCols(local_size) = marginalization_info->linearized_jacobians.middleCols(idx, local_size);
+      }
+    }
+  }
+  return true;
 }
