@@ -9,13 +9,13 @@
 
 class ProjectionFactor : public ceres::SizedCostFunction<2, 7, 7, 7, 1>
 {
-  public:
-    ProjectionFactor(const Eigen::Vector3d &_pts_i, const Eigen::Vector3d &_pts_j);
-    virtual bool Evaluate(double const *const *parameters, double *residuals, double **jacobians) const;
-    void check(double **parameters);
+ public:
+   ProjectionFactor(Eigen::Vector3d  _pts_i, Eigen::Vector3d  _pts_j);
+   bool Evaluate(double const* const* parameters, double* residuals, double** jacobians) const override;
+   void check(double** parameters);
 
-    Eigen::Vector3d pts_i, pts_j;
-    Eigen::Matrix<double, 2, 3> tangent_base;
-    static Eigen::Matrix2d sqrt_info;
-    static double sum_t;
+   Eigen::Vector3d pts_i, pts_j;
+   Eigen::Matrix<double, 2, 3> tangent_base;
+   static Eigen::Matrix2d sqrt_info;
+   static double sum_t;
 };
