@@ -17,86 +17,85 @@ using namespace Eigen;
 
 class FeaturePerFrame
 {
-  public:
-    FeaturePerFrame(const Eigen::Matrix<double, 7, 1> &_point, double td)
-    {
-        point.x() = _point(0);
-        point.y() = _point(1);
-        point.z() = _point(2);
-        uv.x() = _point(3);
-        uv.y() = _point(4);
-        velocity.x() = _point(5); 
-        velocity.y() = _point(6); 
-        cur_td = td;
-    }
-    double cur_td;
-    Vector3d point;
-    Vector2d uv;
-    Vector2d velocity;
-    double z;
-    bool is_used;
-    double parallax;
-    MatrixXd A;
-    VectorXd b;
-    double dep_gradient;
+ public:
+   FeaturePerFrame(const Eigen::Matrix<double, 7, 1>& _point, double td)
+   {
+     point.x() = _point(0);
+     point.y() = _point(1);
+     point.z() = _point(2);
+     uv.x() = _point(3);
+     uv.y() = _point(4);
+     velocity.x() = _point(5);
+     velocity.y() = _point(6);
+     cur_td = td;
+   }
+   double cur_td;
+   Vector3d point;
+   Vector2d uv;
+   Vector2d velocity;
+   double z;
+   bool is_used;
+   double parallax;
+   MatrixXd A;
+   VectorXd b;
+   double dep_gradient;
 };
 
 class FeaturePerId
 {
-  public:
-    const int feature_id;
-    int start_frame;
-    vector<FeaturePerFrame> feature_per_frame;
+ public:
+   const int feature_id; // 特征点的id号
+   int start_frame; // 滑窗内第一次观测到该特征点的帧号
+   vector<FeaturePerFrame> feature_per_frame;
 
-    int used_num;
-    bool is_outlier;
-    bool is_margin;
-    double estimated_depth;
-    int solve_flag; // 0 haven't solve yet; 1 solve succ; 2 solve fail;
+   int used_num;
+   bool is_outlier;
+   bool is_margin;
+   double estimated_depth; // 估计的深度
+   int solve_flag; // 0 haven't solve yet; 1 solve succ; 2 solve fail;
 
-    Vector3d gt_p;
+   Vector3d gt_p;
 
-    FeaturePerId(int _feature_id, int _start_frame)
+   FeaturePerId(int _feature_id, int _start_frame)
         : feature_id(_feature_id), start_frame(_start_frame),
           used_num(0), estimated_depth(-1.0), solve_flag(0)
-    {
-    }
+   { }
 
-    int endFrame();
+   int endFrame();
 };
 
 class FeatureManager
 {
-  public:
-    FeatureManager(Matrix3d _Rs[]);
+ public:
+   explicit FeatureManager(Matrix3d _Rs[]);
 
-    void setRic(Matrix3d _ric[]);
+   void setRic(Matrix3d _ric[]);
 
-    void clearState();
+   void clearState();
 
-    int getFeatureCount();
+   int getFeatureCount();
 
-    bool addFeatureCheckParallax(int frame_count, const map<int, vector<pair<int, Eigen::Matrix<double, 7, 1>>>> &image, double td);
-    void debugShow();
-    vector<pair<Vector3d, Vector3d>> getCorresponding(int frame_count_l, int frame_count_r);
+   bool addFeatureCheckParallax(int frame_count, const map<int, vector<pair<int, Eigen::Matrix<double, 7, 1>>>>& image, double td);
+   void debugShow();
+   vector<pair<Vector3d, Vector3d>> getCorresponding(int frame_count_l, int frame_count_r);
 
-    //void updateDepth(const VectorXd &x);
-    void setDepth(const VectorXd &x);
-    void removeFailures();
-    void clearDepth(const VectorXd &x);
-    VectorXd getDepthVector();
-    void triangulate(Vector3d Ps[], Vector3d tic[], Matrix3d ric[]);
-    void removeBackShiftDepth(Eigen::Matrix3d marg_R, Eigen::Vector3d marg_P, Eigen::Matrix3d new_R, Eigen::Vector3d new_P);
-    void removeBack();
-    void removeFront(int frame_count);
-    void removeOutlier();
-    list<FeaturePerId> feature;
-    int last_track_num;
+   //void updateDepth(const VectorXd &x);
+   void setDepth(const VectorXd& x);
+   void removeFailures();
+   void clearDepth(const VectorXd& x);
+   VectorXd getDepthVector();
+   void triangulate(Vector3d Ps[], Vector3d tic[], Matrix3d ric[]);
+   void removeBackShiftDepth(const Eigen::Matrix3d& marg_R, const Eigen::Vector3d& marg_P, const Eigen::Matrix3d& new_R, const Eigen::Vector3d& new_P);
+   void removeBack();
+   void removeFront(int frame_count);
+   void removeOutlier();
+   list<FeaturePerId> feature;
+   int last_track_num;
 
-  private:
-    double compensatedParallax2(const FeaturePerId &it_per_id, int frame_count);
-    const Matrix3d *Rs;
-    Matrix3d ric[NUM_OF_CAM];
+ private:
+   static double compensatedParallax2(const FeaturePerId& it_per_id, int frame_count);
+   const Matrix3d* Rs;
+   Matrix3d ric[NUM_OF_CAM]; // 相机坐标系到IMU坐标系的外参
 };
 
 #endif
