@@ -1,207 +1,231 @@
 #include "initial_alignment.h"
 
-void solveGyroscopeBias(map<double, ImageFrame> &all_image_frame, Vector3d* Bgs)
+/**
+ * @brief 求解陀螺仪的零偏，同时利用求出来的重新进行预积分
+ * @param[in] all_image_frame
+ * @param[in] Bgs
+ */
+void solveGyroscopeBias(map<double, ImageFrame>& all_image_frame, Vector3d* Bgs)
 {
-    Matrix3d A;
-    Vector3d b;
-    Vector3d delta_bg;
-    A.setZero();
-    b.setZero();
-    map<double, ImageFrame>::iterator frame_i;
-    map<double, ImageFrame>::iterator frame_j;
-    for (frame_i = all_image_frame.begin(); next(frame_i) != all_image_frame.end(); frame_i++)
-    {
-        frame_j = next(frame_i);
-        MatrixXd tmp_A(3, 3);
-        tmp_A.setZero();
-        VectorXd tmp_b(3);
-        tmp_b.setZero();
-        Eigen::Quaterniond q_ij(frame_i->second.R.transpose() * frame_j->second.R);
-        tmp_A = frame_j->second.pre_integration->jacobian.template block<3, 3>(O_R, O_BG);
-        tmp_b = 2 * (frame_j->second.pre_integration->delta_q.inverse() * q_ij).vec();
-        A += tmp_A.transpose() * tmp_A;
-        b += tmp_A.transpose() * tmp_b;
+  Matrix3d A;
+  Vector3d b;
+  Vector3d delta_bg;
+  A.setZero();
+  b.setZero();
+  map<double, ImageFrame>::iterator frame_i;
+  map<double, ImageFrame>::iterator frame_j;
+  for (frame_i = all_image_frame.begin(); next(frame_i) != all_image_frame.end(); frame_i++)
+  {
+    frame_j = next(frame_i);
+    MatrixXd tmp_A(3, 3);
+    tmp_A.setZero();
+    VectorXd tmp_b(3);
+    tmp_b.setZero();                       // 相邻两帧之间的旋转矩阵
+    Eigen::Quaterniond q_ij(frame_i->second.R.transpose() * frame_j->second.R); // 视觉得到的相邻两帧之间的旋转
+    tmp_A = frame_j->second.pre_integration->jacobian.template block<3, 3>(O_R, O_BG); // imu得到的角度误差关于陀螺仪零偏的雅克比
+    tmp_b = 2 * (frame_j->second.pre_integration->delta_q.inverse() * q_ij).vec(); // .vec()是四元数的虚部
+    A += tmp_A.transpose() * tmp_A;
+    b += tmp_A.transpose() * tmp_b;
 
-    }
-    delta_bg = A.ldlt().solve(b);
-    ROS_WARN_STREAM("gyroscope bias initial calibration " << delta_bg.transpose());
+  }
+  delta_bg = A.ldlt().solve(b); // 对称正定矩阵的Cholesky分解
+  ROS_WARN_STREAM("gyroscope bias initial calibration " << delta_bg.transpose());
 
-    for (int i = 0; i <= WINDOW_SIZE; i++)
-        Bgs[i] += delta_bg;
+  // 滑窗中的零偏设置为新求解出来的零偏
+  for (int i = 0; i <= WINDOW_SIZE; i++)
+    Bgs[i] += delta_bg;
 
-    for (frame_i = all_image_frame.begin(); next(frame_i) != all_image_frame.end( ); frame_i++)
-    {
-        frame_j = next(frame_i);
-        frame_j->second.pre_integration->repropagate(Vector3d::Zero(), Bgs[0]);
-    }
+  // 重新进行预积分
+  for (frame_i = all_image_frame.begin(); next(frame_i) != all_image_frame.end( ); frame_i++)
+  {
+    frame_j = next(frame_i);
+    frame_j->second.pre_integration->repropagate(Vector3d::Zero(), Bgs[0]);
+  }
 }
 
 
-MatrixXd TangentBasis(Vector3d &g0)
+MatrixXd TangentBasis(Vector3d& g0) // g0是重力向量的初始值（方向是之前求得的，大小使用了先验）
 {
-    Vector3d b, c;
-    Vector3d a = g0.normalized();
-    Vector3d tmp(0, 0, 1);
-    if(a == tmp)
-        tmp << 1, 0, 0;
-    b = (tmp - a * (a.transpose() * tmp)).normalized();
-    c = a.cross(b);
-    MatrixXd bc(3, 2);
-    bc.block<3, 1>(0, 0) = b;
-    bc.block<3, 1>(0, 1) = c;
-    return bc;
+  Vector3d b, c;
+  Vector3d a = g0.normalized();
+  Vector3d tmp(0, 0, 1);
+  if(a == tmp)
+    tmp << 1, 0, 0;
+  b = (tmp - a * (a.transpose() * tmp)).normalized(); // 得到与a垂直的单位向量
+  c = a.cross(b); // a x b
+  MatrixXd bc(3, 2);
+  bc.block<3, 1>(0, 0) = b;
+  bc.block<3, 1>(0, 1) = c;
+  return bc;
 }
 
-void RefineGravity(map<double, ImageFrame> &all_image_frame, Vector3d &g, VectorXd &x)
+/**
+ * @brief 重力方向的优化
+ * @param[in] all_image_frame
+ * @param[in] g
+ * @param[in] x
+ */
+void RefineGravity(map<double, ImageFrame>& all_image_frame, Vector3d& g, VectorXd& x)
 {
-    Vector3d g0 = g.normalized() * G.norm();
-    Vector3d lx, ly;
-    //VectorXd x;
-    int all_frame_count = all_image_frame.size();
-    int n_state = all_frame_count * 3 + 2 + 1;
+  Vector3d g0 = g.normalized() * G.norm(); // 重力方向的初始值 * 重力加速度的模（固定的已知值）
+  Vector3d lx, ly; // 重力方向的切线基
+  //VectorXd x;
+  int all_frame_count = (int)all_image_frame.size();
+  int n_state = all_frame_count * 3 + 2 + 1;
 
-    MatrixXd A{n_state, n_state};
-    A.setZero();
-    VectorXd b{n_state};
-    b.setZero();
+  MatrixXd A{n_state, n_state};
+  A.setZero();
+  VectorXd b{n_state};
+  b.setZero();
 
-    map<double, ImageFrame>::iterator frame_i;
-    map<double, ImageFrame>::iterator frame_j;
-    for(int k = 0; k < 4; k++)
-    {
-        MatrixXd lxly(3, 2);
-        lxly = TangentBasis(g0);
-        int i = 0;
-        for (frame_i = all_image_frame.begin(); next(frame_i) != all_image_frame.end(); frame_i++, i++)
-        {
-            frame_j = next(frame_i);
-
-            MatrixXd tmp_A(6, 9);
-            tmp_A.setZero();
-            VectorXd tmp_b(6);
-            tmp_b.setZero();
-
-            double dt = frame_j->second.pre_integration->sum_dt;
-
-
-            tmp_A.block<3, 3>(0, 0) = -dt * Matrix3d::Identity();
-            tmp_A.block<3, 2>(0, 6) = frame_i->second.R.transpose() * dt * dt / 2 * Matrix3d::Identity() * lxly;
-            tmp_A.block<3, 1>(0, 8) = frame_i->second.R.transpose() * (frame_j->second.T - frame_i->second.T) / 100.0;     
-            tmp_b.block<3, 1>(0, 0) = frame_j->second.pre_integration->delta_p + frame_i->second.R.transpose() * frame_j->second.R * TIC[0] - TIC[0] - frame_i->second.R.transpose() * dt * dt / 2 * g0;
-
-            tmp_A.block<3, 3>(3, 0) = -Matrix3d::Identity();
-            tmp_A.block<3, 3>(3, 3) = frame_i->second.R.transpose() * frame_j->second.R;
-            tmp_A.block<3, 2>(3, 6) = frame_i->second.R.transpose() * dt * Matrix3d::Identity() * lxly;
-            tmp_b.block<3, 1>(3, 0) = frame_j->second.pre_integration->delta_v - frame_i->second.R.transpose() * dt * Matrix3d::Identity() * g0;
-
-
-            Matrix<double, 6, 6> cov_inv = Matrix<double, 6, 6>::Zero();
-            //cov.block<6, 6>(0, 0) = IMU_cov[i + 1];
-            //MatrixXd cov_inv = cov.inverse();
-            cov_inv.setIdentity();
-
-            MatrixXd r_A = tmp_A.transpose() * cov_inv * tmp_A;
-            VectorXd r_b = tmp_A.transpose() * cov_inv * tmp_b;
-
-            A.block<6, 6>(i * 3, i * 3) += r_A.topLeftCorner<6, 6>();
-            b.segment<6>(i * 3) += r_b.head<6>();
-
-            A.bottomRightCorner<3, 3>() += r_A.bottomRightCorner<3, 3>();
-            b.tail<3>() += r_b.tail<3>();
-
-            A.block<6, 3>(i * 3, n_state - 3) += r_A.topRightCorner<6, 3>();
-            A.block<3, 6>(n_state - 3, i * 3) += r_A.bottomLeftCorner<3, 6>();
-        }
-            A = A * 1000.0;
-            b = b * 1000.0;
-            x = A.ldlt().solve(b);
-            VectorXd dg = x.segment<2>(n_state - 3);
-            g0 = (g0 + lxly * dg).normalized() * G.norm();
-            //double s = x(n_state - 1);
-    }   
-    g = g0;
-}
-
-bool LinearAlignment(map<double, ImageFrame> &all_image_frame, Vector3d &g, VectorXd &x)
-{
-    int all_frame_count = all_image_frame.size();
-    int n_state = all_frame_count * 3 + 3 + 1;
-
-    MatrixXd A{n_state, n_state};
-    A.setZero();
-    VectorXd b{n_state};
-    b.setZero();
-
-    map<double, ImageFrame>::iterator frame_i;
-    map<double, ImageFrame>::iterator frame_j;
+  map<double, ImageFrame>::iterator frame_i;
+  map<double, ImageFrame>::iterator frame_j;
+  for(int k = 0; k < 4; k++)
+  {
+    MatrixXd lxly(3, 2);
+    lxly = TangentBasis(g0); // 重力方向的切线基组成的3x2的矩阵
     int i = 0;
     for (frame_i = all_image_frame.begin(); next(frame_i) != all_image_frame.end(); frame_i++, i++)
     {
-        frame_j = next(frame_i);
+      frame_j = next(frame_i);
 
-        MatrixXd tmp_A(6, 10);
-        tmp_A.setZero();
-        VectorXd tmp_b(6);
-        tmp_b.setZero();
+      MatrixXd tmp_A(6, 9);
+      tmp_A.setZero();
+      VectorXd tmp_b(6);
+      tmp_b.setZero();
 
-        double dt = frame_j->second.pre_integration->sum_dt;
+      double dt = frame_j->second.pre_integration->sum_dt;
 
-        tmp_A.block<3, 3>(0, 0) = -dt * Matrix3d::Identity();
-        tmp_A.block<3, 3>(0, 6) = frame_i->second.R.transpose() * dt * dt / 2 * Matrix3d::Identity();
-        tmp_A.block<3, 1>(0, 9) = frame_i->second.R.transpose() * (frame_j->second.T - frame_i->second.T) / 100.0;     
-        tmp_b.block<3, 1>(0, 0) = frame_j->second.pre_integration->delta_p + frame_i->second.R.transpose() * frame_j->second.R * TIC[0] - TIC[0];
-        //cout << "delta_p   " << frame_j->second.pre_integration->delta_p.transpose() << endl;
-        tmp_A.block<3, 3>(3, 0) = -Matrix3d::Identity();
-        tmp_A.block<3, 3>(3, 3) = frame_i->second.R.transpose() * frame_j->second.R;
-        tmp_A.block<3, 3>(3, 6) = frame_i->second.R.transpose() * dt * Matrix3d::Identity();
-        tmp_b.block<3, 1>(3, 0) = frame_j->second.pre_integration->delta_v;
-        //cout << "delta_v   " << frame_j->second.pre_integration->delta_v.transpose() << endl;
 
-        Matrix<double, 6, 6> cov_inv = Matrix<double, 6, 6>::Zero();
-        //cov.block<6, 6>(0, 0) = IMU_cov[i + 1];
-        //MatrixXd cov_inv = cov.inverse();
-        cov_inv.setIdentity();
+      tmp_A.block<3, 3>(0, 0) = -dt * Matrix3d::Identity();
+      tmp_A.block<3, 2>(0, 6) = 0.5 * frame_i->second.R.transpose() * dt * dt * lxly; //
+      tmp_A.block<3, 1>(0, 8) = frame_i->second.R.transpose() * (frame_j->second.T - frame_i->second.T) / 100.0;
+      tmp_b.block<3, 1>(0, 0) = frame_j->second.pre_integration->delta_p + frame_i->second.R.transpose() * frame_j->second.R * TIC[0] - TIC[0] - frame_i->second.R.transpose() * dt * dt / 2 * g0;
 
-        MatrixXd r_A = tmp_A.transpose() * cov_inv * tmp_A;
-        VectorXd r_b = tmp_A.transpose() * cov_inv * tmp_b;
+      tmp_A.block<3, 3>(3, 0) = -Matrix3d::Identity();
+      tmp_A.block<3, 3>(3, 3) = frame_i->second.R.transpose() * frame_j->second.R;
+      tmp_A.block<3, 2>(3, 6) = frame_i->second.R.transpose() * dt * lxly;
+      tmp_b.block<3, 1>(3, 0) = frame_j->second.pre_integration->delta_v - frame_i->second.R.transpose() * dt * Matrix3d::Identity() * g0;
 
-        A.block<6, 6>(i * 3, i * 3) += r_A.topLeftCorner<6, 6>();
-        b.segment<6>(i * 3) += r_b.head<6>();
 
-        A.bottomRightCorner<4, 4>() += r_A.bottomRightCorner<4, 4>();
-        b.tail<4>() += r_b.tail<4>();
+      Matrix<double, 6, 6> cov_inv = Matrix<double, 6, 6>::Zero();
+      //cov.block<6, 6>(0, 0) = IMU_cov[i + 1];
+      //MatrixXd cov_inv = cov.inverse();
+      cov_inv.setIdentity();
 
-        A.block<6, 4>(i * 3, n_state - 4) += r_A.topRightCorner<6, 4>();
-        A.block<4, 6>(n_state - 4, i * 3) += r_A.bottomLeftCorner<4, 6>();
+      MatrixXd r_A = tmp_A.transpose() * cov_inv * tmp_A;
+      VectorXd r_b = tmp_A.transpose() * cov_inv * tmp_b;
+
+      A.block<6, 6>(i * 3, i * 3) += r_A.topLeftCorner<6, 6>();
+      b.segment<6>(i * 3) += r_b.head<6>();
+
+      A.bottomRightCorner<3, 3>() += r_A.bottomRightCorner<3, 3>();
+      b.tail<3>() += r_b.tail<3>();
+
+      A.block<6, 3>(i * 3, n_state - 3) += r_A.topRightCorner<6, 3>();
+      A.block<3, 6>(n_state - 3, i * 3) += r_A.bottomLeftCorner<3, 6>();
     }
     A = A * 1000.0;
     b = b * 1000.0;
     x = A.ldlt().solve(b);
-    double s = x(n_state - 1) / 100.0;
-    ROS_DEBUG("estimated scale: %f", s);
-    g = x.segment<3>(n_state - 4);
-    ROS_DEBUG_STREAM(" result g     " << g.norm() << " " << g.transpose());
-    if(fabs(g.norm() - G.norm()) > 1.0 || s < 0)
-    {
-        return false;
-    }
-
-    RefineGravity(all_image_frame, g, x);
-    s = (x.tail<1>())(0) / 100.0;
-    (x.tail<1>())(0) = s;
-    ROS_DEBUG_STREAM(" refine     " << g.norm() << " " << g.transpose());
-    if(s < 0.0 )
-        return false;   
-    else
-        return true;
+    VectorXd dg = x.segment<2>(n_state - 3);
+    g0 = (g0 + lxly * dg).normalized() * G.norm();
+    //double s = x(n_state - 1);
+  }
+  g = g0;
 }
 
-bool VisualIMUAlignment(map<double, ImageFrame> &all_image_frame, Vector3d* Bgs, Vector3d &g, VectorXd &x)
+/**
+ * @brief 求解各帧的速度，枢纽帧的重力方向，以及尺度
+ * @param[in] all_image_frame
+ * @param[in] g
+ * @param[in] x
+ * @return bool
+ */
+bool LinearAlignment(map<double, ImageFrame>& all_image_frame, Vector3d& g, VectorXd& x)
 {
-    solveGyroscopeBias(all_image_frame, Bgs);
+  int all_frame_count = (int)all_image_frame.size();
+  int n_state = all_frame_count * 3 + 3 + 1; // 所有帧的速度，枢纽帧的重力方向，以及尺度等总的状态量
 
-    if(LinearAlignment(all_image_frame, g, x))
-        return true;
-    else 
-        return false;
+  MatrixXd A{n_state, n_state};
+  A.setZero();
+  VectorXd b{n_state};
+  b.setZero();
+
+  map<double, ImageFrame>::iterator frame_i;
+  map<double, ImageFrame>::iterator frame_j;
+  int i = 0;
+  for (frame_i = all_image_frame.begin(); next(frame_i) != all_image_frame.end(); frame_i++, i++)
+  {
+    frame_j = next(frame_i);
+
+    MatrixXd tmp_A(6, 10);
+    tmp_A.setZero();
+    VectorXd tmp_b(6);
+    tmp_b.setZero();
+
+    double dt = frame_j->second.pre_integration->sum_dt;
+
+    tmp_A.block<3, 3>(0, 0) = -dt * Matrix3d::Identity();
+    tmp_A.block<3, 3>(0, 6) = 0.5 * frame_i->second.R.transpose() * dt * dt;
+    tmp_A.block<3, 1>(0, 9) = frame_i->second.R.transpose() * (frame_j->second.T - frame_i->second.T) / 100.0;
+    tmp_b.block<3, 1>(0, 0) = frame_j->second.pre_integration->delta_p + frame_i->second.R.transpose() * frame_j->second.R * TIC[0] - TIC[0];
+    //cout << "delta_p   " << frame_j->second.pre_integration->delta_p.transpose() << endl;
+    tmp_A.block<3, 3>(3, 0) = -Matrix3d::Identity();
+    tmp_A.block<3, 3>(3, 3) = frame_i->second.R.transpose() * frame_j->second.R;
+    tmp_A.block<3, 3>(3, 6) = frame_i->second.R.transpose() * dt;
+    tmp_b.block<3, 1>(3, 0) = frame_j->second.pre_integration->delta_v;
+    //cout << "delta_v   " << frame_j->second.pre_integration->delta_v.transpose() << endl;
+
+    Matrix<double, 6, 6> cov_inv = Matrix<double, 6, 6>::Zero(); // 协方差矩阵的逆，也就是信息矩阵（各个量的权重）
+    //cov.block<6, 6>(0, 0) = IMU_cov[i + 1];
+    //MatrixXd cov_inv = cov.inverse();
+    cov_inv.setIdentity();
+
+    MatrixXd r_A = tmp_A.transpose() * cov_inv * tmp_A; // 转换为对称正定矩阵
+    VectorXd r_b = tmp_A.transpose() * cov_inv * tmp_b;
+
+    // 将小的矩阵加到大的矩阵中
+    A.block<6, 6>(i * 3, i * 3) += r_A.topLeftCorner<6, 6>();
+    b.segment<6>(i * 3) += r_b.head<6>(); // 从第i*3个元素开始，取6个元素
+
+    A.bottomRightCorner<4, 4>() += r_A.bottomRightCorner<4, 4>();
+    b.tail<4>() += r_b.tail<4>();
+
+    A.block<6, 4>(i * 3, n_state - 4) += r_A.topRightCorner<6, 4>();
+    A.block<4, 6>(n_state - 4, i * 3) += r_A.bottomLeftCorner<4, 6>();
+  }
+  // 增强数值稳定性
+  A = A * 1000.0;
+  b = b * 1000.0;
+  x = A.ldlt().solve(b);
+  double s = x(n_state - 1) / 100.0;
+  ROS_DEBUG("estimated scale: %f", s);
+  g = x.segment<3>(n_state - 4);
+  ROS_DEBUG_STREAM(" result g     " << g.norm() << " " << g.transpose());
+  if(fabs(g.norm() - G.norm()) > 1.0 || s < 0)
+  {
+    return false;
+  }
+
+  // 重力方向优化
+  RefineGravity(all_image_frame, g, x);
+  // 得到真实的尺度
+  s = (x.tail<1>())(0) / 100.0;
+  (x.tail<1>())(0) = s;
+  ROS_DEBUG_STREAM(" refine     " << g.norm() << " " << g.transpose());
+  if(s < 0.0 )
+    return false;
+  else
+    return true;
+}
+
+bool VisualIMUAlignment(map<double, ImageFrame>& all_image_frame, Vector3d* Bgs, Vector3d& g, VectorXd& x)
+{
+  solveGyroscopeBias(all_image_frame, Bgs);
+
+  if(LinearAlignment(all_image_frame, g, x))
+    return true;
+  else
+    return false;
 }
