@@ -9,7 +9,7 @@
 #include <fstream>
 
 const double FOCAL_LENGTH = 460.0;
-const int WINDOW_SIZE = 10;
+const int WINDOW_SIZE = 10; // 滑窗大小
 const int NUM_OF_CAM = 1;
 const int NUM_OF_F = 1000;
 //#define UNIT_SPHERE_ERROR
@@ -39,28 +39,28 @@ extern int ROLLING_SHUTTER;
 extern double ROW, COL;
 
 
-void readParameters(ros::NodeHandle &n);
+void readParameters(ros::NodeHandle& n);
 
 enum SIZE_PARAMETERIZATION
 {
-    SIZE_POSE = 7,
-    SIZE_SPEEDBIAS = 9,
-    SIZE_FEATURE = 1
+  SIZE_POSE = 7,
+  SIZE_SPEEDBIAS = 9,
+  SIZE_FEATURE = 1
 };
 
 enum StateOrder
 {
-    O_P = 0,
-    O_R = 3,
-    O_V = 6,
-    O_BA = 9,
-    O_BG = 12
+  O_P = 0,
+  O_R = 3,
+  O_V = 6,
+  O_BA = 9,
+  O_BG = 12
 };
 
 enum NoiseOrder
 {
-    O_AN = 0,
-    O_GN = 3,
-    O_AW = 6,
-    O_GW = 9
+  O_AN = 0,
+  O_GN = 3,
+  O_AW = 6,
+  O_GW = 9
 };
