@@ -20,87 +20,87 @@ CameraPoseVisualization keyframebasevisual(0.0, 0.0, 1.0, 1.0);
 static double sum_of_path = 0;
 static Vector3d last_path(0.0, 0.0, 0.0);
 
-void registerPub(ros::NodeHandle &n)
+void registerPub(ros::NodeHandle& n)
 {
-    pub_latest_odometry = n.advertise<nav_msgs::Odometry>("imu_propagate", 1000);
-    pub_path = n.advertise<nav_msgs::Path>("path", 1000);
-    pub_relo_path = n.advertise<nav_msgs::Path>("relocalization_path", 1000);
-    pub_odometry = n.advertise<nav_msgs::Odometry>("odometry", 1000);
-    pub_point_cloud = n.advertise<sensor_msgs::PointCloud>("point_cloud", 1000);
-    pub_margin_cloud = n.advertise<sensor_msgs::PointCloud>("history_cloud", 1000);
-    pub_key_poses = n.advertise<visualization_msgs::Marker>("key_poses", 1000);
-    pub_camera_pose = n.advertise<nav_msgs::Odometry>("camera_pose", 1000);
-    pub_camera_pose_visual = n.advertise<visualization_msgs::MarkerArray>("camera_pose_visual", 1000);
-    pub_keyframe_pose = n.advertise<nav_msgs::Odometry>("keyframe_pose", 1000);
-    pub_keyframe_point = n.advertise<sensor_msgs::PointCloud>("keyframe_point", 1000);
-    pub_extrinsic = n.advertise<nav_msgs::Odometry>("extrinsic", 1000);
-    pub_relo_relative_pose=  n.advertise<nav_msgs::Odometry>("relo_relative_pose", 1000);
+  pub_latest_odometry = n.advertise<nav_msgs::Odometry>("imu_propagate", 1000); //
+  pub_path = n.advertise<nav_msgs::Path>("path", 1000);
+  pub_relo_path = n.advertise<nav_msgs::Path>("relocalization_path", 1000);
+  pub_odometry = n.advertise<nav_msgs::Odometry>("odometry", 1000);
+  pub_point_cloud = n.advertise<sensor_msgs::PointCloud>("point_cloud", 1000);
+  pub_margin_cloud = n.advertise<sensor_msgs::PointCloud>("history_cloud", 1000);
+  pub_key_poses = n.advertise<visualization_msgs::Marker>("key_poses", 1000);
+  pub_camera_pose = n.advertise<nav_msgs::Odometry>("camera_pose", 1000);
+  pub_camera_pose_visual = n.advertise<visualization_msgs::MarkerArray>("camera_pose_visual", 1000);
+  pub_keyframe_pose = n.advertise<nav_msgs::Odometry>("keyframe_pose", 1000);
+  pub_keyframe_point = n.advertise<sensor_msgs::PointCloud>("keyframe_point", 1000);
+  pub_extrinsic = n.advertise<nav_msgs::Odometry>("extrinsic", 1000);
+  pub_relo_relative_pose=  n.advertise<nav_msgs::Odometry>("relo_relative_pose", 1000);
 
-    cameraposevisual.setScale(1);
-    cameraposevisual.setLineWidth(0.05);
-    keyframebasevisual.setScale(0.1);
-    keyframebasevisual.setLineWidth(0.01);
+  cameraposevisual.setScale(1);
+  cameraposevisual.setLineWidth(0.05);
+  keyframebasevisual.setScale(0.1);
+  keyframebasevisual.setLineWidth(0.01);
 }
 
-void pubLatestOdometry(const Eigen::Vector3d &P, const Eigen::Quaterniond &Q, const Eigen::Vector3d &V, const std_msgs::Header &header)
+void pubLatestOdometry(const Eigen::Vector3d& P, const Eigen::Quaterniond& Q, const Eigen::Vector3d& V, const std_msgs::Header& header)
 {
-    Eigen::Quaterniond quadrotor_Q = Q ;
+  Eigen::Quaterniond quadrotor_Q = Q ;
 
-    nav_msgs::Odometry odometry;
-    odometry.header = header;
-    odometry.header.frame_id = "world";
-    odometry.pose.pose.position.x = P.x();
-    odometry.pose.pose.position.y = P.y();
-    odometry.pose.pose.position.z = P.z();
-    odometry.pose.pose.orientation.x = quadrotor_Q.x();
-    odometry.pose.pose.orientation.y = quadrotor_Q.y();
-    odometry.pose.pose.orientation.z = quadrotor_Q.z();
-    odometry.pose.pose.orientation.w = quadrotor_Q.w();
-    odometry.twist.twist.linear.x = V.x();
-    odometry.twist.twist.linear.y = V.y();
-    odometry.twist.twist.linear.z = V.z();
-    pub_latest_odometry.publish(odometry);
+  nav_msgs::Odometry odometry;
+  odometry.header = header;
+  odometry.header.frame_id = "world";
+  odometry.pose.pose.position.x = P.x();
+  odometry.pose.pose.position.y = P.y();
+  odometry.pose.pose.position.z = P.z();
+  odometry.pose.pose.orientation.x = quadrotor_Q.x();
+  odometry.pose.pose.orientation.y = quadrotor_Q.y();
+  odometry.pose.pose.orientation.z = quadrotor_Q.z();
+  odometry.pose.pose.orientation.w = quadrotor_Q.w();
+  odometry.twist.twist.linear.x = V.x();
+  odometry.twist.twist.linear.y = V.y();
+  odometry.twist.twist.linear.z = V.z();
+  pub_latest_odometry.publish(odometry);
 }
 
 void printStatistics(const Estimator &estimator, double t)
 {
-    if (estimator.solver_flag != Estimator::SolverFlag::NON_LINEAR)
-        return;
-    printf("position: %f, %f, %f\r", estimator.Ps[WINDOW_SIZE].x(), estimator.Ps[WINDOW_SIZE].y(), estimator.Ps[WINDOW_SIZE].z());
-    ROS_DEBUG_STREAM("position: " << estimator.Ps[WINDOW_SIZE].transpose());
-    ROS_DEBUG_STREAM("orientation: " << estimator.Vs[WINDOW_SIZE].transpose());
-    for (int i = 0; i < NUM_OF_CAM; i++)
+  if (estimator.solver_flag != Estimator::SolverFlag::NON_LINEAR)
+    return;
+  printf("position: %f, %f, %f\r", estimator.Ps[WINDOW_SIZE].x(), estimator.Ps[WINDOW_SIZE].y(), estimator.Ps[WINDOW_SIZE].z());
+  ROS_DEBUG_STREAM("position: " << estimator.Ps[WINDOW_SIZE].transpose());
+  ROS_DEBUG_STREAM("orientation: " << estimator.Vs[WINDOW_SIZE].transpose());
+  for (int i = 0; i < NUM_OF_CAM; i++)
+  {
+    //ROS_DEBUG("calibration result for camera %d", i);
+    ROS_DEBUG_STREAM("extrinsic tic: " << estimator.tic[i].transpose());
+    ROS_DEBUG_STREAM("extrinsic ric: " << Utility::R2ypr(estimator.ric[i]).transpose());
+    if (ESTIMATE_EXTRINSIC)
     {
-        //ROS_DEBUG("calibration result for camera %d", i);
-        ROS_DEBUG_STREAM("extirnsic tic: " << estimator.tic[i].transpose());
-        ROS_DEBUG_STREAM("extrinsic ric: " << Utility::R2ypr(estimator.ric[i]).transpose());
-        if (ESTIMATE_EXTRINSIC)
-        {
-            cv::FileStorage fs(EX_CALIB_RESULT_PATH, cv::FileStorage::WRITE);
-            Eigen::Matrix3d eigen_R;
-            Eigen::Vector3d eigen_T;
-            eigen_R = estimator.ric[i];
-            eigen_T = estimator.tic[i];
-            cv::Mat cv_R, cv_T;
-            cv::eigen2cv(eigen_R, cv_R);
-            cv::eigen2cv(eigen_T, cv_T);
-            fs << "extrinsicRotation" << cv_R << "extrinsicTranslation" << cv_T;
-            fs.release();
-        }
+      cv::FileStorage fs(EX_CALIB_RESULT_PATH, cv::FileStorage::WRITE);
+      Eigen::Matrix3d eigen_R;
+      Eigen::Vector3d eigen_T;
+      eigen_R = estimator.ric[i];
+      eigen_T = estimator.tic[i];
+      cv::Mat cv_R, cv_T;
+      cv::eigen2cv(eigen_R, cv_R);
+      cv::eigen2cv(eigen_T, cv_T);
+      fs << "extrinsicRotation" << cv_R << "extrinsicTranslation" << cv_T;
+      fs.release();
     }
+  }
 
-    static double sum_of_time = 0;
-    static int sum_of_calculation = 0;
-    sum_of_time += t;
-    sum_of_calculation++;
-    ROS_DEBUG("vo solver costs: %f ms", t);
-    ROS_DEBUG("average of time %f ms", sum_of_time / sum_of_calculation);
+  static double sum_of_time = 0;
+  static int sum_of_calculation = 0;
+  sum_of_time += t;
+  sum_of_calculation++;
+  ROS_DEBUG("vo solver costs: %f ms", t);
+  ROS_DEBUG("average of time %f ms", sum_of_time / sum_of_calculation);
 
-    sum_of_path += (estimator.Ps[WINDOW_SIZE] - last_path).norm();
-    last_path = estimator.Ps[WINDOW_SIZE];
-    ROS_DEBUG("sum of path %f", sum_of_path);
-    if (ESTIMATE_TD)
-        ROS_INFO("td %f", estimator.td);
+  sum_of_path += (estimator.Ps[WINDOW_SIZE] - last_path).norm();
+  last_path = estimator.Ps[WINDOW_SIZE];
+  ROS_DEBUG("sum of path %f", sum_of_path);
+  if (ESTIMATE_TD)
+      ROS_INFO("td %f", estimator.td);
 }
 
 void pubOdometry(const Estimator &estimator, const std_msgs::Header &header)
@@ -403,20 +403,23 @@ void pubKeyframe(const Estimator &estimator)
     }
 }
 
-void pubRelocalization(const Estimator &estimator)
+void pubRelocalization(const Estimator& estimator)
 {
-    nav_msgs::Odometry odometry;
-    odometry.header.stamp = ros::Time(estimator.relo_frame_stamp);
-    odometry.header.frame_id = "world";
-    odometry.pose.pose.position.x = estimator.relo_relative_t.x();
-    odometry.pose.pose.position.y = estimator.relo_relative_t.y();
-    odometry.pose.pose.position.z = estimator.relo_relative_t.z();
-    odometry.pose.pose.orientation.x = estimator.relo_relative_q.x();
-    odometry.pose.pose.orientation.y = estimator.relo_relative_q.y();
-    odometry.pose.pose.orientation.z = estimator.relo_relative_q.z();
-    odometry.pose.pose.orientation.w = estimator.relo_relative_q.w();
-    odometry.twist.twist.linear.x = estimator.relo_relative_yaw;
-    odometry.twist.twist.linear.y = estimator.relo_frame_index;
+  nav_msgs::Odometry odometry;
+  odometry.header.stamp = ros::Time(estimator.relo_frame_stamp);
+  odometry.header.frame_id = "world";
+  // 优化后的回环位姿
+  odometry.pose.pose.position.x = estimator.relo_relative_t.x();
+  odometry.pose.pose.position.y = estimator.relo_relative_t.y();
+  odometry.pose.pose.position.z = estimator.relo_relative_t.z();
+  odometry.pose.pose.orientation.x = estimator.relo_relative_q.x();
+  odometry.pose.pose.orientation.y = estimator.relo_relative_q.y();
+  odometry.pose.pose.orientation.z = estimator.relo_relative_q.z();
+  odometry.pose.pose.orientation.w = estimator.relo_relative_q.w();
+  // 回环帧的yaw
+  odometry.twist.twist.linear.x = estimator.relo_relative_yaw;
+  // 回环帧对应的当前帧在回换节点的index
+  odometry.twist.twist.linear.y = estimator.relo_frame_index;
 
-    pub_relo_relative_pose.publish(odometry);
+  pub_relo_relative_pose.publish(odometry);
 }
