@@ -29,24 +29,24 @@ extern nav_msgs::Path path;
 extern ros::Publisher pub_pose_graph;
 extern int IMAGE_ROW, IMAGE_COL;
 
-void registerPub(ros::NodeHandle &n);
+void registerPub(ros::NodeHandle& n);
 
-void pubLatestOdometry(const Eigen::Vector3d &P, const Eigen::Quaterniond &Q, const Eigen::Vector3d &V, const std_msgs::Header &header);
+void pubLatestOdometry(const Eigen::Vector3d& P, const Eigen::Quaterniond& Q, const Eigen::Vector3d& V, const std_msgs::Header& header);
 
-void printStatistics(const Estimator &estimator, double t);
+void printStatistics(const Estimator& estimator, double t);
 
-void pubOdometry(const Estimator &estimator, const std_msgs::Header &header);
+void pubOdometry(const Estimator& estimator, const std_msgs::Header& header);
 
-void pubInitialGuess(const Estimator &estimator, const std_msgs::Header &header);
+void pubInitialGuess(const Estimator& estimator, const std_msgs::Header& header);
 
-void pubKeyPoses(const Estimator &estimator, const std_msgs::Header &header);
+void pubKeyPoses(const Estimator& estimator, const std_msgs::Header& header);
 
-void pubCameraPose(const Estimator &estimator, const std_msgs::Header &header);
+void pubCameraPose(const Estimator& estimator, const std_msgs::Header& header);
 
-void pubPointCloud(const Estimator &estimator, const std_msgs::Header &header);
+void pubPointCloud(const Estimator& estimator, const std_msgs::Header& header);
 
-void pubTF(const Estimator &estimator, const std_msgs::Header &header);
+void pubTF(const Estimator& estimator, const std_msgs::Header& header);
 
-void pubKeyframe(const Estimator &estimator);
+void pubKeyframe(const Estimator& estimator);
 
-void pubRelocalization(const Estimator &estimator);
+void pubRelocalization(const Estimator& estimator);
